Celery Workers updated - Fixed #12
